split up long test
